astria/pulid_pipeline/requirements.txt
diffusers
torch
torchvision
diffusers
transformers
opencv-python
httpx==0.23.3
timm
einops
ftfy
facexlib
insightface
onnxruntime
onnxruntime-gpu
accelerate
SentencePiece
safetensors
torchsde
xformers
apex
optimum-quanto
accelerate
